Submitted. Needs fixes.
Comments: Delete comments that no longer make sense, for example all the TODO comments. word_counts_free() needs to free all its nodes, not just the enclosing struct. Program crashes with out of memory error. I think the issue is on line 168, sizeof(word) gives the size of the char* data type which is probably 8 bytes, whereas you need to allocate enough bytes for all the characters in the string, which can be found with strlen(word)+1.